move the storage of plural rules to po
This makes sure that the po files are the single source
of truth for translations as well as plural rules. The
translator can update the rules as well. And for the language
names, Intl.DisplayNames can be used and is more versatile.
Then the only thing that needs to be specified in the config
is the codes of the locales, nothing else. This makes the config
simpler.

```js
export default {
    otherLocales: ['es', 'fr'],
    adapters: ...
}
```